refactor: 공용 모듈 common 디렉토리로 파일 이동
