+++ src/common/message.js
