+++ src/common/validate.js
